+++ inst/include/depth.h
@@ -43,7 +43,7 @@ struct node {
 class phylo_tree {
 public:
 
-  phylo_tree(const std::vector< long >& tree_edge) {
+  phylo_tree(const std::vector< int >& tree_edge) {
 
     int root_no = 2 + static_cast<int>(0.25 * tree_edge.size()); // this holds always.
     tree.resize(tree_edge.size() / 2);
@@ -183,7 +183,7 @@ struct node {
 
 class depth_tracker {
 public :
-  depth_tracker(const std::vector< long >& tree_edge) {
+  depth_tracker(const std::vector< int >& tree_edge) {
 
     root_no = 2 + static_cast<int>(0.25 * tree_edge.size()); // this holds always.
     auto max_num = *std::max_element(tree_edge.begin(), tree_edge.end());
